Basic multi machine is now base tile for multimachine
