+++ src/main/java/muramasa/antimatter/tile/TileEntityMachine.java
@@ -30,6 +30,7 @@ import muramasa.antimatter.network.packets.TileGuiEventPacket;
 import muramasa.antimatter.recipe.Recipe;
 import muramasa.antimatter.structure.StructureCache;
 import muramasa.antimatter.texture.Texture;
+import muramasa.antimatter.tile.multi.TileEntityBasicMultiMachine;
 import muramasa.antimatter.tile.multi.TileEntityMultiMachine;
 import muramasa.antimatter.tool.AntimatterToolType;
 import muramasa.antimatter.util.Utils;
@@ -301,7 +302,7 @@ public class TileEntityMachine extends TileEntityTickable implements INamedConta
         coverHandler.ifPresent(machineCoverHandler -> builder.withInitial(AntimatterProperties.MACHINE_TILE, this));
         BlockPos cPos = StructureCache.get(this.getWorld(), pos);
         if (cPos != null) {
-            TileEntityMultiMachine mTile = (TileEntityMultiMachine) world.getTileEntity(cPos);
+            TileEntityBasicMultiMachine mTile = (TileEntityBasicMultiMachine) world.getTileEntity(cPos);
             builder.withInitial(AntimatterProperties.MULTI_MACHINE_TEXTURE,a -> {
                 Texture[] tex = mTile.getMachineType().getBaseTexture(mTile.getMachineTier());
                 if (tex.length == 1) return tex[0];
